nasm: fix issue with repeat configure
This Fixes bug: [YOCTO #2176]

This fixes the following issue as mentioned in the bug:
If nasm-native has to be rebuilt (due to a signature change) it will fail:

| checking for a BSD-compatible install... /usr/bin/install -c
| /media/large/tmp/work/x86_64-linux/nasm-native-2.07-r1/nasm-2.07/configure:
line 4261: syntax error near unexpected token `-W'
| /media/large/tmp/work/x86_64-linux/nasm-native-2.07-r1/nasm-2.07/configure:
line 4261: `PA_ADD_CFLAGS(-W)'
| ERROR: oe_runconf failed
NOTE: package nasm-native-2.07-r1: task do_configure: Failed

The failure appears to be caused by do_patch_fixaclocal being run a second time

+++ meta/recipes-devtools/nasm/nasm_2.07.bb
@@ -13,11 +13,11 @@ SRC_URI[sha256sum] = "ac70ee451c73d742c9ff830502e5f8b1f648b2abffa8fd00944243283b
 
 inherit autotools
 
-do_patch_fixaclocal () {
-	mv aclocal.m4 acinclude.m4
+do_configure_prepend () {
+	if [ -f aclocal.m4 ] && [ ! -f acinclude.m4 ]; then
+		mv aclocal.m4 acinclude.m4
+	fi
 }
-
-addtask do_patch_fixaclocal after do_patch before do_configure
 
 do_install() {
 	install -d ${D}${bindir}
